+++ src/library/fileio/SaveFileList.h
@@ -21,6 +21,7 @@
 #define LIBTAS_SAVEFILELIST_H_INCLUDED
 
 #include <cstdio> // FILE
+#include <string>
 
 namespace libtas {
 
@@ -59,6 +60,9 @@ int getSaveFileFd(const char *file);
 
 /* Get if savefile was removed */
 bool isSaveFileRemoved(const char *file);
+
+/* Get the n-th save file inside directory `dir`. Returns empty string if not present */
+std::string getSaveFileInsideDir(std::string dir, int n);
 
 }